refactor(livechat): move ChatContainer handlers into the connector
Lift the remaining handlers and derived flags (getRoom, onSubmit, onUpload,
typing helpers, handleConnectingAgentAlert, handleQueueMessage,
checkConnectingAgent, can* flags, etc.) into the connector as stable
callbacks, passed to the container as props. The container keeps its class
lifecycle (componentDidMount/Update/Unmount) and just orchestrates them and
renders the presentational component.
